requirements.txt
streamlit
pandas
openpyxl
Pillow
gspread
google-auth
requests